// src/app/interfaces/Register.ts
export interface Register {
    id?: number;
    nomeescola: string;
    responsavel: string;
    email: string;
    telefone: string;
    rua: string;
    numero: string;
    bairro: string;
    cidade: string;
    estado: string;
    cep: string;
    created_at?: string;
    updated_at?: string;
}